requirements.txt
streamlit 
google-generativeai 
langchain 
chromadb 
python-pptx 
PyPDF2 
python-docx